pyproject.toml
[tool.poetry]
name = "cognite-neat"
version = "0.92.3"
readme = "README.md"
description = "Knowledge graph transformation"
authors = [
    "Nikola Vasiljevic <nikola.vasiljevic@cognite.com>",
    "Anders Albert <anders.albert@cognite.com>",
    "Aleksandrs Livincovs <aleksandrs.livincovs@cognite.com>",
]
license = "Apache-2.0"
documentation = "https://cognite-neat.readthedocs-hosted.com/"
homepage = "https://cognite-neat.readthedocs-hosted.com/"
repository = "https://github.com/cognitedata/neat"
packages = [{ include = "cognite", from = "." }]
exclude = [
    "cognite/neat/app/ui/neat-app/public/**",
    "cognite/neat/app/ui/neat-app/src/**",
    "cognite/neat/app/ui/neat-app/node_modules/**",
    "cognite/neat/app/ui/neat-app/*.json",
]

[tool.ruff.lint]
# See https://docs.astral.sh/ruff/rules/ for an overview of ruff rules
select = ["E", "W", "F", "I", "RUF", "TID", "UP", "B", "FLY", "PTH", "ERA"]
fixable = ["E", "W", "F", "I", "RUF", "TID", "UP", "B", "FLY", "PTH", "ERA"]
ignore = []

[tool.ruff.lint.isort]
known-third-party = ["cognite.client"]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = ["scripts"]
fix = true


[tool.mypy]
explicit_package_bases = true
allow_redefinition = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.poetry.scripts]
neat = "cognite.neat._app.main:run"

[tool.poetry.dependencies]
python = "^3.10"
# Core dependencies
pandas = "*"
cognite-sdk = "^7.54.6"
rdflib = "*"
pydantic = "^2"
PyYAML = "*"
requests = "*"
urllib3 = "^2"
openpyxl = "*"
# Remove when legacy is removed
deepdiff = "*"

# Notebook
rich = { version = "^13.7.1", extras = ["jupyter"], optional = true }

# Backport from Python 3.11
exceptiongroup = { version = "^1.1.3", python = "<3.11" }
'backports.strenum' = { version = "^1.2", python = "<3.11" }
typing_extensions = { version = "^4.8", python = "<3.11" }
tomli = { version = "^2.0.1", python = "<3.11" }

# API
uvicorn = { extras = ["standard"], version = "^0", optional = true }
prometheus-client = { version = "^0", optional = true }
fastapi = { version = "^0", optional = true }
python-multipart = { version = "^0", optional = true }

# WorkFlow
schedule = { version = "^1", optional = true }

# Alternative graph for NEAT Store
oxrdflib = { version = "^0.3.3", optional = true, extras = ["oxigraph"] }
pyoxigraph = { version = "0.3.19", optional = true }

# Only used for GoogleSheetImporter
gspread = { version = "*", optional = true }
google-api-python-client = { version = "*", optional = true }
google-auth-oauthlib = { version = "*", optional = true }

# Remove when ledacy is removed
jinja2 = { version = "^3.1.2", optional = true }

# Documentation dependencies
mkdocs = { version = "*", optional = true }
mkdocs-jupyter = { version = "*", optional = true }
mkdocs-material-extensions = { version = "*", optional = true }
mkdocs-git-revision-date-localized-plugin = { version = "*", optional = true }
mkdocs-git-authors-plugin = { version = "*", optional = true }
mkdocs-gitbook = { version = "*", optional = true }
mkdocs-glightbox = { version = "*", optional = true }
pymdown-extensions = { version = "*", optional = true }
mkdocs-autorefs = { version = "^0.5.0", optional = true }
mkdocstrings = { version = "*", optional = true, extras = ["python"] }

[tool.poetry.extras]
docs = [
    "mkdocs",
    "mkdocs-jupyter",
    "mkdocs-material-extensions",
    "mkdocs-git-revision-date-localized-plugin",
    "mkdocs-git-authors-plugin",
    "mkdocs-gitbook",
    "mkdocs-glightbox",
    "pymdown-extensions",
    "mkdocstrings",
    "mkdocs-autorefs",
]
google = ["gspread", "google-api-python-client", "google-auth-oauthlib"]
oxi = ["oxrdflib", "pyoxigraph"]
service = ["uvicorn", "prometheus-client", "fastapi", "schedule", "python-multipart"]
graphql = ["jinja2"]
jupyter = ["rich"]

all = [
    "jinja2",
    "pyoxigraph",
    "oxrdflib",
    "uvicorn",
    "prometheus-client",
    "fastapi",
    "schedule",
    "python-multipart",
]

[tool.poetry.dev-dependencies]
twine = "*"
# Today, 4. March 2024, the latest verson of pytest 8.1 (released yesterday) is not compatible with pytest-regressions
# This limit should be removed when pytest-regressions is updated to be compatible with pytest 8.1 and beoynd.
pytest = { version = ">=8.0, <8.1" }
pytest-cov = "*"
pytest-notebook = "*"
pytest-freezegun = "0.4.2"
pytest-xdist = "^3.6.1"
pre-commit = "*"
jupyter = "*"
pytest-regressions = "*"
memray = [
    { version = "^1.7", platform = "darwin" },
    { version = "^1.7", platform = "linux" },
]
httpx = "*"
mypy = "^1.4"
pandas-stubs = "^2"
types-PyYAML = "^6"
types-openpyxl = "^3"
types-requests = "^2"
types-setuptools = "^68"
cognite-toolkit = "^0.2.4"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
